+++ OUDS/Core/Components/Sources/Dialogs/Alert/Internal/AlertLeadingIcon.swift
@@ -15,6 +15,7 @@ import OUDSTokensSemantic
 import SwiftUI
 
 // MARK: - Alert Leading Icon
+
 struct AlertLeadingIcon: View {
 
     // MARK: Properties